services/payment_service/requirements.txt
# Payment Service dependencies

# Core
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2
python-dotenv==1.0.1
python-decouple==3.8

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Security / crypto / auth
cryptography==42.0.5
pyjwt==2.8.0

# HTTP / tooling
requests==2.31.0
PyYAML==6.0.1

# API docs / observability (optional but pre-pinned)
drf-spectacular==0.27.2
django-prometheus==2.3.1
sentry-sdk==2.0.1

# Async / Message Queue
celery==5.3.6
redis==5.0.3

# Shared libs (local editable install)
-e ../../shared_libs/python
